requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.36
pymysql==1.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
bcrypt==4.0.1

# Validation & Settings
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Development
python-dotenv==1.0.1

# Rate Limiting
slowapi==0.1.9
